% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRegionSubsetSums.R
\name{calcRegionSubsetSums}
\alias{calcRegionSubsetSums}
\title{Calculate Sums for Region Subsets}
\usage{
calcRegionSubsetSums(data, regionSubsetList)
}
\arguments{
\item{data}{A \code{\link[magclass:magclass]{MAgPIE}} object.}

\item{regionSubsetList}{A list of region subsets to calculate of the form
\verb{list(subset_name = c(region_A, region_B, ...)}}
}
\value{
A \code{\link[magclass:magclass]{MAgPIE}} object.
}
\description{
Sum up values in \code{data} for all sets of regions in \code{regionSubsetList}.
}
\examples{
calcRegionSubsetSums(population_magpie, list(xAM = c('LAM', 'NAM')))
}
\author{
Michaja Pehl
}
