% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTimeSeriesStats.R
\name{calcTimeSeriesStats}
\alias{calcTimeSeriesStats}
\title{Calculate Statistics of Time Series}
\usage{
calcTimeSeriesStats(data, var, stats, from = -Inf, to = 2100)
}
\arguments{
\item{data}{A quitte object.}

\item{var}{character(n). The names of the variables to calculate the
statistics on.}

\item{stats}{A named list of functions with two arguments. Each function will
be called with the vector of values as the first argument and the vector of
corresponding periods as the second argument.}

\item{from, to}{Optional variables to filter the time series by period.}
}
\value{
A data frame with columns model, scenario, region, variable, value,
  statistic, unit. "value" contains the calculated statistics. "statistic" is
  name of the statistics, i.e., the name of the function in the argument
  \code{stats}. The "unit" is the unit of the original variable and not
  necessarily the correct unit for the calculated statistics.
}
\description{
Take a quitte object and calculate given summary statistics of the time
series of given variables in the quitte object. Return the result for each
model, scenario, region, variable, and statistic in a data frame.
}
\examples{
  \dontrun{
    calcTimeSeriesStats(
      data, \
      "Emi|CO2",
      stats = list("Net 0 Period" = \(v, p) p[which(v <= 0)[1]]))
  }
}
