% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPiamTemplates.R
\name{checkPiamTemplates}
\alias{checkPiamTemplates}
\title{checkPiamTemplates}
\usage{
checkPiamTemplates(computedVariables, source)
}
\arguments{
\item{computedVariables}{list of reported REMIND variables}

\item{source}{of the reported variables (could be a gdx or folder path),
only used for generating a warning message}
}
\description{
Checks if a given list of reported REMIND variables contains all the variables
expected in the piamInterfaces package.
}
