% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSummations.R
\name{checkSummations}
\alias{checkSummations}
\title{Perform summation check on a dataset}
\usage{
checkSummations(x)
}
\arguments{
\item{x}{a quitte object containing the data}
}
\value{
a list with a message and the failed variable groups as elements
}
\description{
Based on the |+| notation in variable names
}
\examples{

\dontrun{
checkSummations(x)
}

}
\author{
Anastasis Giannousakis
}
