% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVariableNames.R
\name{checkVariableNames}
\alias{checkVariableNames}
\title{checkVariablesNames checks remind2 reporting on inconsistency in variable names}
\usage{
checkVariableNames(vars)
}
\arguments{
\item{vars}{vector with variable names and units such as "PE (EJ)"}
}
\description{
Pass a vector of variable names including the units. Get warnings if inconsistencies are found
for the reporting
}
\author{
Oliver Richters
}
