% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVsCalibData.R
\name{checkVsCalibData}
\alias{checkVsCalibData}
\title{Render checkVsCalibData}
\usage{
checkVsCalibData(
  gdx,
  outputDir = getwd(),
  outputFile = "Check_vs_CalibData.pdf"
)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{outputDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}

\item{outputFile}{\code{character(1)}. File name (without extension) of the
output document to be created.}
}
\value{
The value returned by \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Render checkVsCalibData
}
\examples{
\dontrun{
# Simple use. Creates PDF:
checkVsCalibData(
  gdx = "path/to/data.gdx",
  outputDir = "path/to/output/directory",
  outputFile = "Check_vs_CalibData_Example.pdf"
)
}
}
\author{
Falk Benke
}
