% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorScenConf.R
\name{colorScenConf}
\alias{colorScenConf}
\title{take scenario-config*.csv files and produce colorful xlsx file with values
from main.gms in first row, identical values in this column in turquoise,
and unknown column names in red}
\usage{
colorScenConf(fileList = "", remindPath = ".", expanddata = FALSE)
}
\arguments{
\item{fileList}{vector containing csv file paths}

\item{remindPath}{path to REMIND directory containing the main.gms}

\item{expanddata}{fill all the data based on copyConfigFrom and main.gms}
}
\value{
nothing. For each file in fileList a _colorful.xlsx file is written.
}
\description{
take scenario-config*.csv files and produce colorful xlsx file with values
from main.gms in first row, identical values in this column in turquoise,
and unknown column names in red
}
\examples{

 \dontrun{
    colorScenConf(fileList = c("scenario_config.csv"), remindPath = ".")
  }

}
\author{
Oliver Richters
}
