% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenarios2.R
\name{compareScenarios2}
\alias{compareScenarios2}
\title{Render CompareScenarios2}
\usage{
compareScenarios2(...)
}
\arguments{
\item{...}{parameters to be passed on to piamPlotComparison::compareScenarios}
}
\value{
The value returned by \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
A deprecated function. Please refer to piamPlotComparison::compareScenarios
}
