% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenConf.R
\name{compareScenConf}
\alias{compareScenConf}
\title{take two REMIND scenario-config*.csv files and print the difference,
comparing it to a default.cfg
Can also be used for piamInterfaces mapping files with row.names=NULL and expanddata=FALSE.}
\usage{
compareScenConf(
  fileList = NULL,
  remindPath = "/p/projects/rd3mod/github/repos/remindmodel/remind/develop",
  row.names = 1,
  renamedCols = NULL,
  renamedRows = NULL,
  printit = TRUE,
  expanddata = TRUE
)
}
\arguments{
\item{fileList}{vector containing one csv file paths or two paths as c(oldfile, newfile)
If one, searches same filename in defaultPath. If NULL, user can select}

\item{remindPath}{path to REMIND directory containing main.gms. Used for extracting the
default configuration and find a file for comparison if only one config file is supplied in fileList}

\item{row.names}{column in csv used for row.names. Defaults to 1.
Use NULL for mapping files which either picks the 'variable' column or uses the first two columns.}

\item{renamedCols}{vector with old and new column names such as c("old1" = "new1", "old2" = "new2"))}

\item{renamedRows}{vector with old and new row names such as c("old3" = "new3", "old3" = "new4", "old5" = "new5"))
the "old" name can also remain in the new file, if you generated a variant}

\item{printit}{boolean switch (default: TRUE) whether function prints its output}

\item{expanddata}{fill empty cells with default values. Use FALSE for mapping files}
}
\value{
list with $allwarnings and $out
}
\description{
take two REMIND scenario-config*.csv files and print the difference,
comparing it to a default.cfg
Can also be used for piamInterfaces mapping files with row.names=NULL and expanddata=FALSE.
}
\examples{

 \dontrun{
    compareScenConf(fileList = c("scenario_config_old.csv", "scenario_config_new.csv"),
    renamedCols = c("old1" = "new1", "old2" = "new2"),
    renamedRows = c("old3" = "new3", "old4" = "new4"))
  }

}
\author{
Oliver Richters
}
