% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF.R
\name{convGDX2MIF}
\alias{convGDX2MIF}
\title{Read in GDX and write *.mif reporting}
\usage{
convGDX2MIF(
  gdx,
  gdx_ref = NULL,
  file = NULL,
  scenario = "default",
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150),
  gdx_refpolicycost = gdx_ref,
  testthat = FALSE,
  extraData = NULL
)
}
\arguments{
\item{gdx}{a GDX as created by gdx2::readGDX, or the file name of a gdx}

\item{gdx_ref}{reference-gdx for < cm_startyear, used for fixing the prices to this scenario}

\item{file}{name of the mif file which will be written, if no name is
provided a magpie object containing all the reporting information is
returned}

\item{scenario}{scenario name that is used in the *.mif reporting}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}

\item{gdx_refpolicycost}{reference-gdx for policy costs, a GDX as created by gdx2::readGDX, or the file name of a gdx}

\item{testthat}{boolean whether called by tests, turns some messages into warnings}

\item{extraData}{path to extra data files to be used in the reporting}
}
\description{
Read in all information from GDX file and create
the *.mif reporting
}
\examples{
\dontrun{
convGDX2MIF(gdx, gdx_refpolicycost, file = "REMIND_generic_default.csv", scenario = "default")
}

}
\author{
Lavinia Baumstark
}
