% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF_REMIND2MAgPIE.R
\name{convGDX2MIF_REMIND2MAgPIE}
\alias{convGDX2MIF_REMIND2MAgPIE}
\title{Read in GDX and write *.mif short reporting for REMIND-MAgPIE coupling}
\usage{
convGDX2MIF_REMIND2MAgPIE(
  gdx,
  file = NULL,
  scenario = "default",
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150)
)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{file}{name of the mif file which will be written. If no name is
provided a magpie object containing all the reporting information is
returned}

\item{scenario}{scenario name that is used in the *.mif reporting}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}
}
\description{
Read in information from GDX file and create the *.mif reporting
using only reporting functions that calculate the variables that
are relevant for the REMIND-MAgPIE coupling
}
\examples{

\dontrun{convGDX2MIF_REMIND2MAgPIE(gdx,file="REMIND_generic_default.csv",scenario="default")}

}
\author{
David Klein
}
