% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarListHtml.R
\name{createVarListHtml}
\alias{createVarListHtml}
\title{Create an HTML Document of a Hierarchical List of Variables}
\usage{
createVarListHtml(x, outFileName, title = "List of Variables")
}
\arguments{
\item{x}{A character vector of variable names, a quitte object, or a path to
a mif file.}

\item{outFileName}{A single string. The path of the output file, preferably
ending in \code{.html}}

\item{title}{The title displayed at the top of the created HTML.}
}
\description{
Creates a hierarchical list from mif data using \code{\link{variablesAsList}}
and writes it as an HTML document that displays the hierarchy via the
\code{<details>} HTML5-tag.
}
\examples{
\dontrun{
loadModeltest()
createVarListHtml(data, "variables.html")
}
}
