% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePlus.R
\name{deletePlus}
\alias{deletePlus}
\title{Delete "+" in Variable Names. Optionally Save in File with Suffix
"_withoutPlus"}
\usage{
deletePlus(mif, writemif = FALSE)
}
\arguments{
\item{mif}{A path to a mif-file (might be created by confGDX2MIF), a magpie
object containing reporting, a quitte object, or a character vector of
variable names.}

\item{writemif}{A single truth value. Only used when mif is a file path.
Should the new mif with the adjusted names be saved? If yes the new name
will have the suffix "_withoutPlus.mif".}
}
\value{
If mif is a file path and \code{writemif} is \code{FALSE}, a MAgPIE
  object. If mif is not a file path the result is the object \code{mif} with
  replaced variable names.
}
\description{
The function can take a multiple of input formats. It replaces "|+|", "|++|",
"|+++|", ... by "|" in variable names and in this "deletes" the
"plus-notation".
}
\examples{
deletePlus(c(
  "ab|cd", # "ab|cd"
  "de|+|fgh", # "de|fgh"
  "i|j|++|k|l|++++|mno|p", # "i|j|k|l|mno|p"
  "++|q+|+r|+p+|u+v|+", # "++|q+|+r|+p+|u+v|+"
  "w|+|+|x")) # "w|x"
}
