% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAuxiliaryFile.R
\name{downloadAuxiliaryFile}
\alias{downloadAuxiliaryFile}
\title{Helper to download auxiliary file needed for reporting from RSE server}
\usage{
downloadAuxiliaryFile(fname)
}
\arguments{
\item{fname}{name of the file to be downloaded from RSE server folder with extra data}
}
\description{
Helper to download auxiliary file needed for reporting from RSE server
}
\author{
Falk Benke
}
