% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoGAINS2025AirPollutants.R
\name{exoGAINS2025AirPollutants}
\alias{exoGAINS2025AirPollutants}
\title{Calculate projected air pollutant emissions at the level of GAINS sectors
based on REMIND activities, GAINS emission factors, and baseyear (2020)
emissions}
\usage{
exoGAINS2025AirPollutants(
  remind_output,
  gains_emifacs,
  baseyear_emis_2020,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150),
  logging = FALSE
)
}
\arguments{
\item{remind_output}{a magpie object containing all needed REMIND activities}

\item{gains_emifacs}{a magpie object containing historic and projected
GAINS emission factors for all 35 GAINS sectors}

\item{baseyear_emis_2020}{a magpie object containing baseyear (2020)
emissions for all 35 GAINS sectors}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}

\item{logging}{Boolean to activate logging}
}
\description{
Calculate projected air pollutant emissions at the level of GAINS sectors
based on REMIND activities, GAINS emission factors, and baseyear (2020)
emissions
}
\examples{
\dontrun{
exoGAINS2025AirPollutants(remind_output, gains_emifacs, baseyear_emis_2020)
}
}
\author{
Laurin Köhler-Schindler
}
