% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCs2Profiles.R
\name{getCs2Profiles}
\alias{getCs2Profiles}
\title{Get the List of cs2 Profiles}
\usage{
getCs2Profiles(path = "./scripts/cs2/profiles.json")
}
\arguments{
\item{path}{\code{character(1)}. Path to the profiles file.}
}
\value{
A named list of profiles. The names are the names of the respective profile. Each profile is a named list. Those names correspond to arguments of \code{\link{compareScenarios2}}.
}
\description{
A list of arguments to \code{\link{compareScenarios2}} is called a (cs2-) profile. This function loads a list of such profiles from a file. The default argument is set such that the default profiles file of REMIND is loaded.
}
\examples{
\dontrun{
profiles <- getCs2Profiles() # load from default location
profiles <- getCs2Profiles(path = "path/to/profiles.json")
}
}
\author{
Christof Schoetz
}
