% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLine}
\alias{getLine}
\title{gets characters (line) from the terminal or from a connection}
\usage{
getLine()
}
\value{
string
}
\description{
gets characters (line) from the terminal or from a connection
}
