% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPath.R
\name{getMifScenPath}
\alias{getMifScenPath}
\alias{getMifHistPath}
\alias{getCfgScenPath}
\alias{getCfgDefaultPath}
\title{Get Paths to Certain Files in the REMIND Directory}
\usage{
getMifScenPath(outputDirs, mustWork = FALSE)

getMifHistPath(outputDirs, mustWork = FALSE)

getCfgScenPath(outputDirs, mustWork = FALSE)

getCfgDefaultPath(remindDir = ".", mustWork = FALSE)
}
\arguments{
\item{outputDirs}{A character vector of paths to output folders of REMIND runs.}

\item{remindDir}{A single string. The path to the remind directory.}

\item{mustWotk}{\code{logical(1)}. Throw error if file not available?}
}
\description{
\code{getMifScenPath}: get path to the scenarios' reporting mifs.
\code{getMifHistPath}: get path to the scenarios' historical.mif.
\code{getCfgScenPath}: get path to the scenarios' config.Rdata.
\code{getCfgDefaultPath}: get path to REMIND's  default.cfg.
}
