% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModeltest.R
\name{loadCs2Data}
\alias{loadCs2Data}
\title{Load compareScenarios2 Data}
\usage{
loadCs2Data(
  mifScen,
  mifHist,
  cfgScen = NULL,
  cfgDefault = NULL,
  envir = globalenv()
)
}
\arguments{
\item{mifScen}{\code{character(n)}, optionally named. Paths to scenario mifs.
If the vector has names, those are used to refer to the scenarios in the
output file.}

\item{mifHist}{\code{character(1)}. Path to historical mif.}

\item{cfgScen, cfgDefault}{See section "YAML Parameters" in \link[=compareScenarios2]{compareScenarios2()}.}

\item{envir}{\code{environment}. The environment where the loaded data is put into.}
}
\description{
Load data from mif files into R-objects as used in \link[=compareScenarios2]{compareScenarios2()}.
}
\examples{
\dontrun{
loadCs2Data(
  c("path/to/Base.mif", "path/to/NDC.mif"),
  "path/to/historical.mif")
}
}
