% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModeltest.R
\name{loadModeltest}
\alias{loadModeltest}
\title{Load Modeltest Results}
\usage{
loadModeltest(
  envir = globalenv(),
  namePattern = "^SSP2-.*-AMT$",
  folder = tempdir()
)
}
\arguments{
\item{envir}{\code{environment}. The environment where the loaded data is put into.}

\item{namePattern}{\code{character(1)}. A regular expression to filter the
modeltest run names.}

\item{folder}{\code{character(1)}. A folder to copy the modeltest data to.}
}
\description{
The newest model tests are collected from the cluster and copied into a
temporary folder (by default). Then the \link[=compareScenarios]{compareScenarios()}
data loading procedure is used to load this data into the users environment.
}
\examples{
\dontrun{
loadModeltest()

ssp1 <- new.env()
ssp2eu <- new.env()
loadModeltest(ssp1, "^SSP1-AMT-")
loadModeltest(ssp2eu, "^SSP2-.*-AMT$")
ssp1$data
ssp2eu$data
}
}
