% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyInvestmentVariables.R
\name{modifyInvestmentVariables}
\alias{modifyInvestmentVariables}
\title{Modify Investment Variables}
\usage{
modifyInvestmentVariables(x, ref = NULL, startYear = NULL)
}
\arguments{
\item{x}{a magclass object to be manipulated}

\item{ref}{an optional magclass object to be used for fixing values before 'startYear'}

\item{startYear}{years before will be overwritten with values from 'ref'}
}
\description{
This function transforms investment variables to the normal reporting convention.
}
\details{
Years represented by investment variables in the energy system
('vm_deltaCap', 'vm_costInvTeDir' and 'vm_costInvTeAdj') are different from the normal
reporting convention. In the current REMIND version, vm_deltacap(t) represents
the average of the years t-4..t, while in the general reporting
convention it represents the average of t-2.5..t+2.5 (for 5 year time steps).

See also: https://github.com/remindmodel/remind/pull/1238.
}
\author{
Falk Benke
}
