% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashAnalysis.R
\name{nashAnalysis}
\alias{nashAnalysis}
\title{Nash Analysis}
\usage{
nashAnalysis(gdx = "fulldata.gdx", outputFile = NULL)
}
\arguments{
\item{gdx}{file path to a gdx file (default fulldata.gdx)}

\item{outputFile}{file name to save the html dashboard}
}
\description{
Create plots visualizing nash convergence of a given REMIND run
}
\author{
Falk Benke
}
