% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashConvergenceReport.R
\name{nashConvergenceReport}
\alias{nashConvergenceReport}
\title{Nash Convergence Report}
\usage{
nashConvergenceReport(gdx = "fulldata.gdx", outputDir = getwd())
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{outputDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}
}
\description{
Create plots visualizing nash convergence of a given REMIND run
}
\author{
Falk Benke and Renato Rodrigues
}
