% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNashConvergence.R
\name{plotNashConvergence}
\alias{plotNashConvergence}
\title{Creates a REMIND convergence overview}
\usage{
plotNashConvergence(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\description{
Creates a REMIND convergence overview
}
\examples{

  \dontrun{
    plotNashConvergence(gdx="fulldata.gdx")
  }

}
\author{
Renato Rodrigues, Falk Benke
}
