% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAll.R
\name{readAll}
\alias{readAll}
\title{readAll}
\usage{
readAll(pathToGdx, func, asList = TRUE, ...)
}
\arguments{
\item{pathToGdx}{A vector or list. Can contain either the filenames of gdx files or GDX lists as created by readGDX.
If it is named, the names will also be used for the output)}

\item{func}{The output function that should be executed on the gdx files. E.g. emissions}

\item{asList}{If TRUE, the output will be a list of length(gdx). If FALSE, read_all tries to store everything in one
magpie object.}

\item{...}{Additional arguments passed to func.}
}
\value{
A list of magpie objects (as.list=TRUE) or one magpie object (as.list=FALSE) with the output returned by func
        for all the gdx files
}
\description{
(This function was copied from magpie::read_all to remove the magpie dependency.)
Function to read in input from multiple gdx files.
}
\examples{
\dontrun{
gdxPaths <- c(baseline = "fulldata1.gdx", policy = "fulldata2.gdx")
croparea <- read_all(gdxPaths, func = croparea, level = "glo", crop_aggr = TRUE, asList = TRUE)
}
}
\author{
Markus Bonsch
}
