% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind2-package.R
\docType{package}
\name{remind2-package}
\alias{remind2-package}
\alias{remind2}
\title{The REMIND R package (2nd generation)}
\description{
Contains the REMIND-specific routines for data and model output manipulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/remind2}
}

}
\author{
\strong{Maintainer}: Renato Rodrigues \email{renato.rodrigues@pik-potsdam.de}

Authors:
\itemize{
  \item Renato Rodrigues \email{renato.rodrigues@pik-potsdam.de}
  \item Lavinia Baumstark
  \item Falk Benke
  \item David Bantje
  \item Jan Philipp Dietrich
  \item Alois Dirnaichner
  \item Tabea Dorndorf
  \item Jakob Duerrwaechter
  \item Pascal Führlich
  \item Anastasis Giannousakis
  \item Robin Hasse
  \item Jérome Hilaire
  \item David Klein
  \item Johannes Koch
  \item Katarzyna Kowalczyk
  \item Fabrice Lécuyer
  \item Antoine Levesque
  \item Aman Malik
  \item Anne Merfort
  \item Leon Merfort
  \item Simón Morena-Leiva
  \item Adrian Odenweller
  \item Michaja Pehl
  \item Robert Pietzcker
  \item Sebastian Rauner
  \item Oliver Richters
  \item Marianna Rottoli
  \item Tonn Rüter \email{tonn.rueter@pik-potsdam.de}
  \item Robert Salzwedel
  \item Christof Schötz
  \item Felix Schreyer
  \item Kais Siala
  \item Björn Sörgel
  \item Mike Spahr
  \item Jessica Strefler
  \item Philipp Verpoort
  \item Pascal Weigmann
  \item Simon Krogmann
}

}
