% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportClimate.R
\name{reportClimate}
\alias{reportClimate}
\title{Read in GDX and extract climate assessment variables}
\usage{
reportClimate(gdx, output)
}
\arguments{
\item{gdx}{A GDX as created by gdx2::readGDX, or the file name of a gdx}

\item{output}{Existing magclass object containing data generated by other report*.R functions}
}
\description{
Read climate assessment variables from GDX file
}
\examples{
\dontrun{
output <- reportMacroEconomy(gdx, NULL)
reportClimate(gdx, output)
}

}
\author{
Tonn Rüter
}
