% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCrossSubmodels.R
\name{reportCrossSubmodels}
\alias{reportCrossSubmodels}
\title{Read a finished report and calculate variables that need not only variables produced by other
report*.R functions, (from convGDX2MIF.R), but also variables coming from submodels for
which the reporting is done outside of it, such as EDGE-T and MAGICC.
#'}
\usage{
reportCrossSubmodels(inmifpath)
}
\arguments{
\item{inmifpath}{a character string with the path to a REMIND mif file, that
should contain only one model and scenario, and already have both the REMIND and EDGE-T variables}
}
\description{
Read a finished report and calculate variables that need not only variables produced by other
report*.R functions, (from convGDX2MIF.R), but also variables coming from submodels for
which the reporting is done outside of it, such as EDGE-T and MAGICC.
#'
}
\author{
Gabriel Abrahao
}
