% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDIETER.R
\name{reportDIETER}
\alias{reportDIETER}
\title{Reporting for the coupled DIETER Model}
\usage{
reportDIETER(dieterDatafile = "report_DIETER.gdx", outputDir = ".")
}
\arguments{
\item{dieterDatafile}{full path with name of dieter gdx file.}

\item{outputDir}{path to the output folder, default is current folder.}
}
\description{
*Warning* The function modifies the "REMIND_generic_<scenario>.mif" file by appending the
additional reporting variables.
}
\author{
Chen Gong Pratik Agrawal
}
