% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmi.R
\name{reportEmi}
\alias{reportEmi}
\title{Read in GDX and calculate emissions, used in convGDX2MIF.R for the
reporting}
\usage{
reportEmi(
  gdx,
  output = NULL,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150),
  extraData = NULL
)
}
\arguments{
\item{gdx}{a GDX as created by gdx2::readGDX, or the file name of a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}

\item{extraData}{path to extra data files to be used in the reporting}
}
\description{
Read in GDX and calculate emissions, used in convGDX2MIF.R for the
reporting
}
\examples{
\dontrun{
reportEmi(gdx)
}
}
\author{
Felix Schreyer, Anne Merfort
}
