% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmiForClimateAssessment.R
\name{reportEmiForClimateAssessment}
\alias{reportEmiForClimateAssessment}
\title{Reports emissions & air pollutant values from GDX for climate assessment in between Nash iterations before some of
the energy system variables are defined. The report contains only a subset of reported emissions, with the main
difference being that Emi|CA|CO2|Energy and Industrial Processes is calculated from the difference between total and
LUC emissions. Only global values from this function should be used, as it also skips the subtraction
of certain non-regional sources, such as bunkers, from the regional information}
\usage{
reportEmiForClimateAssessment(
  gdx,
  output = NULL,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150),
  extraData = NULL
)
}
\arguments{
\item{gdx}{a GDX as created by gdx2::readGDX, or the file name of a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default: t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}

\item{extraData}{path to extra data files to be used in the reporting}
}
\description{
Reports emissions & air pollutant values from GDX for climate assessment in between Nash iterations before some of
the energy system variables are defined. The report contains only a subset of reported emissions, with the main
difference being that Emi|CA|CO2|Energy and Industrial Processes is calculated from the difference between total and
LUC emissions. Only global values from this function should be used, as it also skips the subtraction
of certain non-regional sources, such as bunkers, from the regional information
}
\examples{
\dontrun{
reportEmiForClimateAssessment(gdx)
}
}
\author{
Gabriel Abrahao, Tonn Rüter
}
