% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFE.R
\name{reportFE}
\alias{reportFE}
\title{Read in GDX and calculate final energy, used in convGDX2MIF.R for the
reporting}
\usage{
reportFE(
  gdx,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150)
)
}
\arguments{
\item{gdx}{a GDX as created by gdx2::readGDX, or the file name of a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}
}
\description{
Read in final energy information from GDX file, information used in
convGDX2MIF.R for the reporting
}
\examples{

  \dontrun{reportFE(gdx)}

}
\author{
Renato Rodrigues, Felix Schreyer
}
