% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCOE.R
\name{reportLCOE}
\alias{reportLCOE}
\title{Read in GDX and calculate LCOE reporting used in convGDX2MIF_LCOE.}
\usage{
reportLCOE(gdx, output.type = "both")
}
\arguments{
\item{gdx}{a GDX object as created by gdx2::readGDX, or the path to a gdx}

\item{output.type}{string to determine which output shall be produced.
Can be either "average" (returns only average LCOE),
"marginal" (returns only marginal LCOE), "both" (returns marginal and average LCOE) and
and "marginal detail" (returns table to trace back how marginal LCOE are calculated).}
}
\value{
MAgPIE object - LCOE calculated by model post-processing.
Two types a) standing system LCOE b) new plant LCOE.
}
\description{
This function provides a post-processing calculation of LCOE (Levelized Cost
of Energy) for energy conversion technologies in REMIND.
It includes most technologies that generate secondary energy and the distribution
technologies which convert secondary energy to final energy.
This script calculates two different types of LCOE: average LCOE (standing system)
and marginal LCOE (new plant).
The average LCOE reflect the total cost incurred by the technology deployment
in a specific time step divided by its energy output.
The marginal LCOE estimate the per-unit lifetime cost of the output if the model
added another capacity of that technology in the respective time step.
The marginal LCOE are calculate in two versions: 1) with fuel prices and co2
taxes of the time step for which the LCOE are calculated (time step prices),
or 2) with an intertemporal weighted-average of fuel price and co2 taxes over
the lifetime of the plant (intertemporal prices).
}
\examples{
\dontrun{
reportLCOE(gdx)
}

}
\author{
Felix Schreyer, Robert Pietzcker, Lavinia Baumstark
}
