% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPrices.R
\name{reportPrices}
\alias{reportPrices}
\title{Read in GDX and calculate prices, used in convGDX2MIF.R for the reporting}
\usage{
reportPrices(
  gdx,
  output = NULL,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150),
  gdx_ref = NULL,
  extraData = NULL
)
}
\arguments{
\item{gdx}{a GDX object as created by gdx2::readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}

\item{gdx_ref}{a GDX object as created by gdx2::readGDX, or the path to a gdx of the reference run.
It is used to guarantee consistency for Moving Avg prices before cm_startyear}

\item{extraData}{path to extra data files to be used in the reporting
(might be needed when reportEmi is executed from within this function)}
}
\value{
MAgPIE object - contains the price variables
}
\description{
Read in price information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{
\dontrun{
reportPrices(gdx)
}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Alois Dirnaichner, Felix Schreyer, David Klein, Renato Rodrigues, Falk Benke
}
