% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEmployment.R
\name{runEmployment}
\alias{runEmployment}
\title{Computes the employment values (jobs) across different sectors}
\usage{
runEmployment(pathToMIF, improvements, multiplier, subtype, shareManf, decline)
}
\arguments{
\item{pathToMIF}{A mif file putput from a REMIND run}

\item{improvements}{Either "None", "CEEW", "Dias", "Rutovitz_aus","Solar_found" or "All". Use "All" for all improvements.}

\item{multiplier}{controls how the regional multiplier for non-oecd countries changes with time.}

\item{subtype}{Subtype for how shares of solar rooftop, wind offshore, and small hydro are assumed in the future. Options "current", "irena", and "expert". See calcDspvShare for more information.}

\item{shareManf}{Either "current" or "local". Current implies current shares of world manufacture remain same until 2050, current means that in 2050 all countries manufacture required components locally.}

\item{decline}{How should the employment factors change over time? "capcosts" means according to capital costs. "static" means it doesn't change}
}
\value{
A magpie object
}
\description{
This function returns a magpie object containing the reporting the jobs for different technologies
}
\examples{
\dontrun{

runEmployment(pathToMIF, improvements = "All", multiplier = "own", subtype = "expert", shareManf = "local", decline = "capcosts")
}
}
\author{
Aman Malik
}
