% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyVariableNames.R
\name{substituePlaceholder}
\alias{substituePlaceholder}
\title{Substitute Placeholders in Variable Names.}
\usage{
substituePlaceholder(x, writeMifSuffix = NULL, substitutions = c(. = "$dot$"))
}
\arguments{
\item{x}{A path to a mif-file (might be created by confGDX2MIF), a quitte
object, or a character vector or factor of variable names.}

\item{writeMifSuffix}{\code{NULL} or a suffix for the filename if the result
should be written as a mif-file.}

\item{substitutions}{A named character vector of placeholders with
replacements as names.}
}
\value{
If \code{x} is not a file path the result is the object \code{x} with
  replaced variable names.
}
\description{
Substitute Placeholders in Variable Names.
}
\examples{
substituePlaceholder(c(
  "3$dot$5", # "3.5"
  "3$dot5", # "3$dot5"
  "3.5", # "3.5"
  "$dot$dot$dot$dot$dot$")) # ".dot.dot."
}
