% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchValuesScenConf.R
\name{switchValuesScenConf}
\alias{switchValuesScenConf}
\title{searches all scenario config files of REMIND directory for a switch and prints the occurences}
\usage{
switchValuesScenConf(switchname = NULL, directory = ".")
}
\arguments{
\item{switchname}{string with switch that is searched in scenario config files.
If NULL, prints all switches with only one value in main.gms and scenario config files}

\item{directory}{path to REMIND directory}
}
\value{
values used in default and all scenario configs
}
\description{
searches all scenario config files of REMIND directory for a switch and prints the occurences
}
\author{
Oliver Richters
}
