% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_cumulate.R
\name{time_cumulate}
\alias{time_cumulate}
\title{time_cumulate}
\usage{
time_cumulate(x, includeEndYear = FALSE)
}
\arguments{
\item{x}{a mapgie object}

\item{includeEndYear}{boolean indicating logic for summing up
when set to TRUE, the value for 2010 is the sum of the values from 2005 - 2010,
otherwise the value for 2010 is the sum of 2005 - 2009}
}
\description{
calculate cumulated values over time for a magpie object
}
\author{
Falk Benke
}
