% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesAsList.R
\name{variablesAsList}
\alias{variablesAsList}
\title{Variable Names as Hierarchical List}
\usage{
variablesAsList(vars, entry = c("NULL", "name", "count", "list"))
}
\arguments{
\item{vars}{A character vector of variable names. Hierarchical section
structure of variables is declared by the symbol \code{|}.}

\item{entry}{A string determining the entries of all leafs and the nodes
which represent variables in \code{vars}. \code{"NULL"} puts \code{NULL}
into the leafs of the resulting list. \code{"name"} places the full name of
an existing variable in the respective node. \code{"count"} shows the
number of occurrences of the respective variable name in \code{vars} as the
first entry of a node. \code{"list"} puts list entries \code{v} and
\code{count} inside each node that represents a variable occurring in
\code{vars} with the full variable name and the number of occurrences,
respectively.}
}
\value{
A hierarchical named list with the structure defined by the use of
  \code{|}.
}
\description{
Take a character vector of variables names structured into groups and
subgroups by \code{|} in the name and convert it into a hierarchical
structure of named lists.
}
\examples{
vars <- c(
  "Emi|GHG|CO2", "Emi|GHG|CH4", "Emi|NOX",
  "FE", "FE|Buildings", "FE|Industry", "FE|Transport")
varList <- variablesAsList(vars)
\dontrun{View(varList)}
varList <- variablesAsList(vars, "count")
\dontrun{View(varList)}
\dontrun{
v <- variablesAsList(data$variable, "list")
mip::showLinePlots(data, v$Emi$GHG$v)}

\dontrun{View(variablesAsList("path/to/scenario.mif"))}
}
\author{
Christof Schoetz
}
