% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSummations.R
\name{checkSummations}
\alias{checkSummations}
\title{Perform summation check on a dataset.}
\usage{
checkSummations(x, tolerance = 1e-05)
}
\arguments{
\item{x}{a quitte object containing the data}

\item{tolerance}{a numeric value of accepted tolerance for parent = sum(children)}
}
\value{
a quitte with the failed and/or missing variable groups (contains a
column "parent" to group by)
}
\description{
Based on the |+| notation in variable names.
}
\examples{
\dontrun{
checkSummations(x)
}

}
\author{
Anastasis Giannousakis
}
