% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorScenConf.R
\name{colorScenConf}
\alias{colorScenConf}
\title{take scenario-config*.csv files and produce colorful xlsx file with values
from configfile in first row, identical values in this column in turquoise,
and unknown column names in red}
\usage{
colorScenConf(fileList = "", configfile = "default.cfg")
}
\arguments{
\item{fileList}{vector containing csv file paths}

\item{configfile}{path to configfile. If empty, uses './default.cfg'}
}
\value{
nothing. For each file in fileList a _colorful.xlsx file is written.
}
\description{
take scenario-config*.csv files and produce colorful xlsx file with values
from configfile in first row, identical values in this column in turquoise,
and unknown column names in red
}
\examples{

 \dontrun{
    colorScenConf(fileList = c("scenario_config.csv"), configfile = "default.cfg")
  }

}
\author{
Oliver Richters
}
