% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenarios2.R
\name{compareScenarios2}
\alias{compareScenarios2}
\title{Render CompareScenarios2}
\usage{
compareScenarios2(
  ...,
  outputDir = getwd(),
  outputFile = "CompareScenarios2",
  outputFormat = "PDF"
)
}
\arguments{
\item{...}{YAML parameters, see below.}

\item{outputDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}

\item{outputFile}{\code{character(1)}. File name (without extension) of the
output document to be created.}

\item{outputFormat}{\code{character(1)}, not case-sensitive. \code{"html"},
\code{"pdf"}, or \code{"rmd"}.}
}
\value{
The value returned by \code{\link[rmarkdown]{rmarkdown::render()}}.
}
\description{
Renders the *.Rmd-files associated to CompareScenarios2. In the Rmds,
scenario- and historical .mif-files are loaded. Then plots are created from
this data. The result may be rendered to PDF or HTML. Alternatively one can
choose Rmd as output format and obtain a copy of the *.Rmd-files.
}
\section{YAML Parameters}{

\describe{
  \item{\code{mifScen}}{
    \code{character(n)}, optionally named. 
    Paths to scenario mifs. If the vector has names, those are used to refer to the scenarios in the output file.}
  \item{\code{mifHist}}{
    \code{character(1)}. 
    Path to historical mif.}
  \item{\code{yearsScen}}{
    \code{numeric(n)}. 
    Default: \code{c(seq(2005, 2060, 5), seq(2070, 2100, 10))}.
    Years to show for scenario data.}
  \item{\code{yearsHist}}{
    \code{numeric(n)}. 
    Default: \code{c(seq(1960, 2020, 1), seq(2025, 2100, 5))}. 
    Years to show for historical data.}
  \item{\code{yearsBarPlot}}{
    \code{numeric(n)}. 
    Default: \code{c(2010, 2030, 2050, 2100)}.
    Years to show in bar plots of scenario data.}
  \item{\code{reg}}{
    \code{NULL} or \code{character(n)}. 
    Default: \code{NULL}.
    Regions to show. \code{NULL} means all.}
  \item{\code{sections}}{
    \code{character(n)}. 
    Default: \code{"all"}.
    Names of sections to include. A subset of  
    \code{c("01_summary", "02_macro", "03_emissions", "04_energy_supply", 
    "05_energy_demand", "06_energy_services", "07_climate", "08_sdp")} 
    or \code{"all"} for all available sections.}
  \item{\code{userSectionPath}}{
    \code{NULL} or \code{character(n)}. 
    Default: \code{NULL}.
    Path to a *.Rmd-file that may be included as additional section.}
  \item{\code{mainReg}}{
    \code{character(1)}. 
    Default: \code{"World"}.
    A region for which larger plots are shown.}
  \item{\code{figWidth, figHeight}}{
    \code{numeric(1)}. 
    Default: \code{14} and \code{8}, respectively.
    Size of plots in inches.}
  \item{\code{warning}}{
    \code{logical(1)}. 
    Default: \code{TRUE}.
    Show warnings in output?}
  \item{\code{dpi}}{
    \code{numeric(1)}. 
    Default: \code{72}.
    Dots per inch of plots.}
}
}

\examples{
\dontrun{
compareScenarios2(
  outputFile = "CompareScenarios2Example1",
  mifScen = c("path/to/Base.mif", "path/to/NDC.mif"),
  mifHist = "path/to/historical.mif",
  userSectionPath = "path/to/myPlots.Rmd")
compareScenarios2(
  outputFile = "CompareScenarios2Example2",
  mifScen = c(ScenarioName1="path/to/scen1.mif", ScenarioName2="path/to/scen2.mif"),
  mifHist = "path/to/historical.mif",
  figWidth = 18, figHeight = 10)
}
}
\author{
Christof Schoetz
}
