% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenConf.R
\name{compareScenConf}
\alias{compareScenConf}
\title{take two REMIND scenario-config*.csv files and print the difference,
comparing it to a default.cfg}
\usage{
compareScenConf(
  fileList = "",
  configfile = "default.cfg",
  row.names = 1,
  renamedCols = NULL,
  renamedRows = NULL,
  printit = TRUE
)
}
\arguments{
\item{fileList}{vector containing two csv file paths. If empty, user can select}

\item{configfile}{path to configfile. If empty, uses './default.cfg'}

\item{row.names}{column in csv used for row.names. Use NULL for mapping files}

\item{renamedCols}{vector with old and new column names such as c("old1" = "new1", "old2" = "new2"))}

\item{renamedRows}{vector with old and new row names such as c("old3" = "new3", "old4" = "new4"))}

\item{printit}{boolean switch (default: TRUE) whether function prints its output}
}
\value{
list with $allwarnings and $out
}
\description{
take two REMIND scenario-config*.csv files and print the difference,
comparing it to a default.cfg
}
\examples{

 \dontrun{
    compareScenConf(fileList = c("scenario_config1.csv", "scenario_config2.csv"),
    configfile = "default.cfg",
    renamedCols = c("old1" = "new1", "old2" = "new2"),
    renamedRows = c("old3" = "new3", "old4" = "new4"))
  }

}
\author{
Oliver Richters
}
