% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarListHtml.R
\name{createVarListHtml}
\alias{createVarListHtml}
\title{Create an HTML Document of a Hierarchical List of Variables}
\usage{
createVarListHtml(
  x,
  outFileName,
  title = "List of Variables",
  usePlus = FALSE,
  details = NULL
)
}
\arguments{
\item{x}{A character vector of variable names, a quitte object, or a
character vectors of paths to mif files.}

\item{outFileName}{A single string. The path of the output file, preferably
ending in \code{.html}}

\item{title}{The title displayed at the top of the created HTML.}

\item{usePlus}{\code{logical(1)}. If \code{FALSE}, removes \code{|+|, |++|,
...} from variable names.}

\item{details}{\code{NULL} or a data frame with a column \code{name}. The
entries of further columns of \code{details} are put into the INFO nodes.}
}
\description{
Creates a hierarchical list from mif data using \code{\link{variablesAsList}}
and writes it as an HTML document that displays the hierarchy via the
\code{<details>} HTML5-tag.
}
\examples{
\dontrun{
loadModeltest()
createVarListHtml(data, "variables.html")
detailsAR6 <-
  readr::read_delim(
    paste0(
      "https://raw.githubusercontent.com/pik-piam/",
      "project_interfaces/master/ar6/mapping_template_AR6.csv"),
    delim = ";",
    col_select = c(r21m42, Definition)
  ) \%>\%
  rename(name = r21m42)
createVarListHtml(
  data,
  "variablesWithDescription.html",
  title = "Reported REMIND Variables with AR6 Description",
  usePlus = TRUE,
  details = detailsAR6)
}
}
\seealso{
\code{\link{variablesAsList}}
}
