% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyInvestmentVariables.R
\name{modifyInvestmentVariables}
\alias{modifyInvestmentVariables}
\title{Modify Investment Variables}
\usage{
modifyInvestmentVariables(x, ref = NULL, startYear = NULL)
}
\arguments{
\item{x}{a magclass object to be manipulated}

\item{ref}{an optional magclass object to be used for fixing values before 'startYear'}

\item{startYear}{years before will be overwritten with values from 'ref'}
}
\description{
a helper to adapt investment variables in reporting after changes in
https://github.com/remindmodel/remind/pull/1238
}
\details{
for a given magclass object, calculates the value of each timestep
as the average of this timestep and the next timestep
the data for the last timestep in the magclass object remains unchanged
}
\author{
Falk Benke
}
