% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashAnalysis.R
\name{nashAnalysis}
\alias{nashAnalysis}
\title{Nash Analysis}
\usage{
nashAnalysis(
  gdx = "fulldata.gdx",
  outputDir = getwd(),
  outputFile = "Nash Analysis.html"
)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{outputDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}

\item{outputFile}{\code{character(1)}. File name (without extension) of the
output document to be created.}
}
\value{
The value returned by \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Create plots visualizing nash convergence of a given REMIND run
}
\author{
Falk Benke
}
