% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind2-package.R
\docType{package}
\name{remind2-package}
\alias{remind2-package}
\alias{remind2}
\title{The REMIND R package (2nd generation)}
\description{
Contains the REMIND-specific routines for data and model output manipulation.
}
