% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAirPollutantEmissions.R
\name{reportAirPollutantEmissions}
\alias{reportAirPollutantEmissions}
\title{Read in GDX, GAINS2025 emission factors and baseyear emissions and
calculate air pollution emissions, used in convGDX2MIF.R for
the reporting}
\usage{
reportAirPollutantEmissions(
  gdx,
  output = NULL,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150),
  extraData = NULL,
  addTmpTotal = FALSE
)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}

\item{extraData}{path to extra data files to be used in the reporting}

\item{addTmpTotal}{choose if temporary totals are computed (default = FALSE)}
}
\description{
Warning: MAgPIE AP emissions are currently not available in REMIND.
         Once available, they need to be reported here, added to the
         temporary total below (see addTmpTotal) and  to the totals
         in reportExtraEmissions.
}
\examples{
\dontrun{
reportAirPollutantEmissions(gdx)
}
}
\author{
Laurin Köhler-Schindler
}
