% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCapacity.R
\name{reportCapacity}
\alias{reportCapacity}
\title{Read in GDX and calculate capacities, used in convGDX2MIF.R for the reporting}
\usage{
reportCapacity(
  gdx,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150)
)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in capacity information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{
\dontrun{
reportCapacity(gdx)
}
}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Lavinia Baumstark, Christoph Bertram
}
