% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportClimate.R
\name{reportClimate}
\alias{reportClimate}
\title{Read in GDX and extract climate assessment variables}
\usage{
reportClimate(gdx)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}
}
\description{
Read climate assessment variables from GDX file
}
\examples{
\dontrun{
reportClimate(gdx)
}

}
\author{
Tonn Rüter
}
