% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCosts.R
\name{reportCosts}
\alias{reportCosts}
\title{Read in GDX and calculate costs, used in convGDX2MIF.R for the reporting}
\usage{
reportCosts(
  gdx,
  output = NULL,
  regionSubsetList = NULL,
  t = c(seq(2005, 2060, 5), seq(2070, 2110, 10), 2130, 2150)
)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}

\item{regionSubsetList}{a list containing regions to create report variables region
aggregations. If NULL (default value) only the global region aggregation "GLO" will
be created.}

\item{t}{temporal resolution of the reporting, default:
t=c(seq(2005,2060,5),seq(2070,2110,10),2130,2150)}
}
\value{
MAgPIE object - contains the cost variables
}
\description{
Read in cost information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportCosts(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
David Klein
}
