% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExtraEmissions.R
\name{reportExtraEmissions}
\alias{reportExtraEmissions}
\title{Report Extra Emissions}
\usage{
reportExtraEmissions(mif, extraData)
}
\arguments{
\item{mif}{a mif file with reported variables from remind2 and EDGE-Transport}

\item{extraData}{path to extra data files to be used in the reporting}
}
\description{
Calulate additonal emissions based on CEDS 2020 emissions and
REMIND/EDGE-Transport Data
}
\author{
Gabriel Abrahão
}
