% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDPVariables.R
\name{reportSDPVariables}
\alias{reportSDPVariables}
\title{Add SDP variables to mif, used in convGDX2MIF.R for the reporting}
\usage{
reportSDPVariables(gdx, output = NULL)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}
}
\value{
MAgPIE object - contains the SDP
}
\description{
Add SDP variables to mif, used in convGDX2MIF.R for the reporting
}
\examples{

\dontrun{reportSDPVariables(output)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Rauner
}
