% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesAsList.R
\name{variablesAsList}
\alias{variablesAsList}
\title{Variable Names as Hierarchical List}
\usage{
variablesAsList(x, entry = c("NULL", "name", "INFO"))
}
\arguments{
\item{x}{A character vector of variable names, a quitte object, or a path to
a mif file.}

\item{entry}{A string determining the entries of all leafs and the nodes
which represent variables in \code{vars}. \code{"NULL"} puts \code{NULL}
into the leafs of the resulting list. \code{"name"} places the full name of
an existing variable as entry \code{nm} in the respective node.
\code{"INFO"} puts a list of further information about this node in each
node with existing variables, which requires \code{x} to be a mif file or a
quitte object.}
}
\value{
A hierarchical named list.
}
\description{
Take a character vector of variables names with hierarchical structure
indicated by \code{|} in the name and convert it into a hierarchical
structure of named lists.
}
\examples{
vars <- c(
  "Emi|GHG|CO2", "Emi|GHG|CH4", "Emi|NOX",
  "FE", "FE|Buildings", "FE|Industry", "FE|Transport")
v <- variablesAsList(vars)
\dontrun{View(v)}
v <- variablesAsList(vars, entry = "name")
\dontrun{mip::showLinePlots(data, v$Emi$GHG$nm)}

\dontrun{
v <- variablesAsList(data, entry = "INFO")
View(v)
}

\dontrun{View(variablesAsList("path/to/scenario.mif"))}
}
\author{
Christof Schoetz
}
